chore(apps/sample-vue-full): fix ducplicate code with sonarcloud issue
