+++ scripts/generators/generate-crud.ts
@@ -25,28 +25,9 @@
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs';
 import { dirname, resolve } from 'node:path';
 import { pathToFileURL } from 'node:url';
+import { parseArgs } from './utils.ts';
 
 // ─── CLI argument parsing ─────────────────────────────────────────────────────
-
-/**
- * Parses a flat `--key value` argument list into a plain object.
- * Consecutive `--key` tokens consume the immediately following token as their value.
- * Unknown or boolean-style flags (no following value) are stored as empty strings.
- *
- * @param argv - The argument list to parse, typically `process.argv.slice(2)`.
- * @returns A map of flag name (without the `--` prefix) to its string value.
- */
-function parseArgs(argv: string[]): Record<string, string> {
-  const result: Record<string, string> = {};
-  for (let i = 0; i < argv.length; i++) {
-    const arg = argv[i];
-    if (arg.startsWith('--')) {
-      result[arg.slice(2)] = argv[i + 1] ?? '';
-      i++;
-    }
-  }
-  return result;
-}
 
 const args = parseArgs(process.argv.slice(2));
 const schemaFilePath = args.schema;
